add code app.js
